requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# API Clients
ccxt>=4.0.0
aiohttp>=3.9.0
requests>=2.31.0
pycoingecko>=3.1.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# LLM Providers
openai>=1.0.0
google-generativeai>=0.3.0

# Discord Integration
discord.py>=2.3.0
aiohttp>=3.9.0

# Scheduling
apscheduler>=3.10.0

# Technical Analysis
ta>=0.10.0
pandas-ta>=0.4.67b0

# Utilities
pydantic>=2.0.0
loguru>=0.7.0
rich>=13.0.0

# Async Support
asyncio-throttle>=1.0.0
